feat: move score to own file
