+++ pong/src/events/mod.rs
@@ -1,4 +1,5 @@
 #![allow(dead_code)]
 
 pub(crate) mod game_events;
+pub(crate) mod score;
 pub(crate) mod reset;
