+++ pong/src/main.rs
@@ -1,6 +1,5 @@
 use internal::components::*;
 use internal::consts::*;
-use internal::player::*;
 use internal::spawn::*;
 
 use events::game_events::*;
@@ -14,29 +13,6 @@ use bevy_rapier2d::prelude::*;
 
 mod events;
 mod internal;
-
-#[derive(Default, Resource)]
-struct Score(HashMap<Player, i32>);
-
-/// Detect if player scored and give
-fn score_run(
-	mut events: EventReader<GameEvents>,
-	mut score_text: Query<(&mut Text, &Player)>,
-	mut scores: ResMut<Score>,
-) {
-	for event in events.read() {
-		if let GameEvents::GainPoint(player) = event {
-			*scores.0.entry(*player).or_default() += 1;
-			let scores = scores.0.get(player).cloned().unwrap_or(0);
-			for (mut text, owner) in &mut score_text {
-				if owner != player {
-					continue;
-				}
-				text.sections[0].value = scores.to_string();
-			}
-		}
-	}
-}
 
 /// Move paddles based on input
 fn move_paddle(
